+++ CHANGELOG.md
@@ -19,9 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     * `.stylelintrc.yml`
     * `fontello.config.json`
 - `<Tag>`, `icBEM()` and `prefixClass()` are now typed by Flow.
-- Add `autoLoadMore` prop in `<InfiniteScroll>`, it will auto trigger `onLoadMore` callback if its height is smaller than 2 times of container's height. (#57)
+- Add `fillSpace` prop in `<InfiniteScroll>`, auto fill spaces with `onLoadMore` callback if its height is smaller than 2 times of container's height. (#57)
 - Remove `disabled` prop in `<InfiniteScroll>`. (#57)
-
 
 
 ## [0.11.1]